Added in working array
I created the single dimensional array, as well as made the board smaller because all that math was just making my head hurt. Now each dimension is stored inside the array in the correct place and there are dead cells inside as well. Will be working on getting the actual game logic to work. This will make my brain hurt even more